# docs/openapi.yaml
openapi: 3.0.3
info:
  title: news-api
  description: |
    REST API backed by the GNews search endpoint.
    When `CLIENT_API_KEYS` is set on the server, send a matching `X-API-Key` header on `/api/*` requests.
  version: 1.2.0
  license:
    name: MIT
    url: https://opensource.org/license/mit
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/v1/articles:
    get:
      summary: Search articles with response metadata
      operationId: searchArticlesV1
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ClientApiKey"
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Count"
        - $ref: "#/components/parameters/Lang"
        - $ref: "#/components/parameters/Country"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/SortBy"
      responses:
        "200":
          description: Enveloped article list from provider (possibly cached)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleSearchEnvelope"
        "400":
          $ref: "#/components/responses/V1BadRequest"
        "401":
          $ref: "#/components/responses/V1Unauthorized"
        "429":
          description: Rate limit exceeded
        "502":
          $ref: "#/components/responses/V1BadGateway"
        "500":
          $ref: "#/components/responses/V1InternalError"
...
  /api/articles:
    get:
      summary: Search articles
      operationId: searchArticles
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - name: X-API-Key
          in: header
          required: false
          schema:
            type: string
          description: Required when the server sets CLIENT_API_KEYS.
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: lang
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-zA-Z]{2}$"
          description: Two-letter article language code.
        - name: country
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-zA-Z]{2}$"
          description: Two-letter source country code.
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include articles published at or after this ISO 8601 timestamp.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Include articles published at or before this ISO 8601 timestamp.
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [publishedAt, relevance]
          description: Provider sort order. Forwarded to GNews as `sortby`.
      responses:
        "200":
          description: Article list from provider (possibly cached)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Rate limit exceeded
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Required for `/api/*` requests only when the server sets `CLIENT_API_KEYS`.
  parameters:
    ClientApiKey:
      name: X-API-Key
      in: header
      required: false
      schema:
        type: string
      description: Required when the server sets CLIENT_API_KEYS.
    Query:
      name: query
      in: query
      required: true
      schema:
        type: string
    Count:
      name: count
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Lang:
      name: lang
      in: query
      required: false
      schema:
        type: string
        pattern: "^[a-zA-Z]{2}$"
      description: Two-letter article language code.
    Country:
      name: country
      in: query
      required: false
      schema:
        type: string
        pattern: "^[a-zA-Z]{2}$"
      description: Two-letter source country code.
    From:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Include articles published at or after this ISO 8601 timestamp.
    To:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      description: Include articles published at or before this ISO 8601 timestamp.
    SortBy:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        enum: [publishedAt, relevance]
      description: Provider sort order. Forwarded to GNews as `sortby`.
  responses:
    Unauthorized:
      description: Missing or invalid X-API-Key when CLIENT_API_KEYS is configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    BadRequest:
      description: Invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    BadGateway:
      description: Upstream provider error or invalid provider payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    V1Unauthorized:
      description: Missing or invalid X-API-Key when CLIENT_API_KEYS is configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    V1BadRequest:
      description: Invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    V1BadGateway:
      description: Upstream provider error or invalid provider payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    V1InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              example: missing_query_parameter
            message:
              type: string
              example: "Missing or empty query parameter: query"
            requestId:
              type: string
...
    ArticleSource:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string
    Article:
      type: object
      required: [title, url, publishedAt, source]
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        url:
          type: string
        image:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
        source:
          $ref: "#/components/schemas/ArticleSource"
    ArticleSearchMeta:
      type: object
      required: [query, count, filters, cache, requestId]
      properties:
        query:
          type: string
        count:
          type: integer
        filters:
          type: object
          additionalProperties: true
        cache:
          type: string
          enum: [hit, miss, coalesced]
        requestId:
          type: string
    ArticleSearchEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        meta:
          $ref: "#/components/schemas/ArticleSearchMeta"
...